libc: Cache large /etc/hosts files
The de-facto standard for ad-blocking on Android is to have a large
hosts database with many 127.0.0.1 entries. Bionic currently does a
linear, line-by-line search in this file for each DNS request, where
each request for a domain that is not blocked needs O(N) string
comparisons. If hosts files were sorted, one could instead do a binary
search in O(log(N)) time.

This commit adds a cache for hosts files that are >=100 KiB large in
/data/misc/net/hosts.cache. The cache uses a list of

 struct _hosts_entry { char ip[40]; char hostname[255]; },

structures, with both still in ASCII format. The list is sorted in
hostnames' strcasecmp() order, such that cache lookups can be performed
using binary search. The cache is validated on each lookup using the
mtime of both the cache and the hosts file. The code is hardened
against race conditions of multiple programs trying to regenerate the
cache simultaneously. If cache lookup fails, this commit falls back to
using the old code.

Updated to Patchset 9

Change-Id: Ie234a94126cdfe8364695406ddba56ab21081da4